+++ crates/module-system/module-implementations/sov-chain-state/tests/integration-tests/dynamic_gas_update.rs
@@ -29,7 +29,7 @@ fn setup_dynamic_gas_update_tests() -> (TestData<S>, TestRunner<TestChainStateRu
 
     let genesis_config =
         HighLevelOptimisticGenesisConfig::generate().add_accounts(vec![TestUser::<S>::generate(
-            (u64::MAX as u128) / 2,
+            u128::MAX / 2,
         )
         .add_token_info(UserTokenInfo {
             token_name: token_name.clone(),
@@ -95,7 +95,7 @@ fn test_gas_price_increases_if_gas_used_exceeds_gas_target() {
                 },
                 mint_to_address: user.address(),
             })
-            .with_max_fee((u64::MAX as u128) / 2),
+            .with_max_fee(Amount::from(u64::MAX / 2)),
         assert: Box::new(move |result, _| {
             assert!(result.tx_receipt.is_successful());
 
@@ -145,7 +145,7 @@ fn test_gas_price_decreases_if_gas_used_is_below_gas_target() {
                     token_id: token_name.id(),
                 },
             })
-            .with_max_fee((u64::MAX as u128) / 2),
+            .with_max_fee(Amount::from(u64::MAX / 2)),
         assert: Box::new(move |result, _| {
             assert!(result.tx_receipt.is_successful());